Moving stuff from main to Game
